Instrucoes atualizadas, DTO´s refatoradas, Service Funcionando em conjunto com Controller ClienteController, envio de email funcionando, e consulta viacep funcionando,

+++ src/main/java/br/org/serratec/dto/ItemPedidoDTO.java
@@ -8,7 +8,7 @@ import lombok.Data;
 @Data
 public class ItemPedidoDTO {
 
-    private Integer idProduto;
+    private Long idProduto;
 	private String nomeProduto;
 
 	@NotNull
